放开验证码校验、修复对冲套利存金宝与账户信息黄金余额对不上（主要是数据库数据影响）

+++ src/main/java/com/fx/xzt/sys/mapper/SaveGoldRecordMapper.java
@@ -48,6 +48,6 @@ public interface SaveGoldRecordMapper extends BaseMapper<SaveGoldRecord> {
 	 * @Description：存金宝买入总金额
 	 * @return
 	 */
-	Float countBuyGold();
+	Double countBuyGold();
 
 }
